+++ libraries/TeenAstroLX200io/src/TeenAstroLX200io.h
@@ -70,10 +70,10 @@ LX200RETURN readMicroLX200(const uint8_t &axis, uint8_t &microStep);
 LX200RETURN writeMicroLX200(const uint8_t &axis, const uint8_t &microStep);
 LX200RETURN readSilentStepLX200(const uint8_t &axis, uint8_t &cool);
 LX200RETURN writeSilentStepLX200(const uint8_t &axis, const uint8_t &cool);
-LX200RETURN readLowCurrLX200(const uint8_t &axis, uint8_t &lowCurr);
-LX200RETURN writeLowCurrLX200(const uint8_t &axis, const uint8_t &lowCurr);
-LX200RETURN readHighCurrLX200(const uint8_t &axis, uint8_t &highCurr);
-LX200RETURN writeHighCurrLX200(const uint8_t &axis, const uint8_t &highCurr);
+LX200RETURN readLowCurrLX200(const uint8_t &axis, unsigned int &lowCurr);
+LX200RETURN writeLowCurrLX200(const uint8_t &axis, const unsigned int &lowCurr);
+LX200RETURN readHighCurrLX200(const uint8_t &axis, unsigned int &highCurr);
+LX200RETURN writeHighCurrLX200(const uint8_t &axis, const unsigned int &highCurr);
 LX200RETURN readFocuserConfig(unsigned int& startPosition, unsigned int& maxPosition,
                               unsigned int& minSpeed, unsigned int& maxSpeed,
                               unsigned int& cmdAcc, unsigned int& manAcc, unsigned int& manDec);
